requirements.txt
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.10
black==25.1.0
certifi==2025.4.26
click==8.2.1
coverage==7.8.2
dill==0.4.0
exceptiongroup==1.3.0
fastapi==0.115.12
flake8==7.2.0
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
isort==6.0.1
Mako==1.3.10
mangum==0.19.0
MarkupSafe==3.0.2
mccabe==0.7.0
mockito==1.5.4
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pydantic==2.11.5
pydantic_core==2.33.2
pyflakes==3.3.2
Pygments==2.19.1
pylint==3.3.7
pytest==8.4.0
pytest-cov==6.1.1
python-decouple==3.8
python-dotenv==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tomli==2.2.1
tomlkit==0.13.3
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
validate_docbr==1.11.0
